Replace renderer lower bound with range display offset
Remove lower bound property from all renderers.
Instead, introduce new display offset in range.
Idea behind this change is, that in most cases,
a renderer's display range is used unshifted, such
that the value can be omitted in most cases.  But
when the display range is shifted, it is typically
specific to the range, not to the renderer, such
that the shift should be specified specifically
for the range, and not for the renderer.