Bugfix: Folders are not editable
When clicking on a folder node, there is no
editor available.  Hence, when clicking on a
folder node, the targeted tree node must be
flagged as non-editable to avoid throwing an
exception within the Swing GUI.  Instead,
when clicking on a folder node, the folder
now will be selected (e.g. for a subsequent
copy / paste / bulk dump operation).